+++ docs/src/markdown/about/changelog.md
@@ -11,6 +11,8 @@
 -   **NEW**: Deprecate non-standard CAM16 (Jab) space. People should use the standard CAM16 JMh or the CAM16 UCS, SCD,
     or LCD Jab spaces. The non-standard Jab is still available via `coloraide.spaces.cam16.CAM16`, but it is no longer
     available in `coloraide.everything` and will be removed at a future time.
+-   **NEW**: Add new channel aliases: `j` for `jz` in Jzazbz and JzCzhz. Also add `c` for `cz` and `h` for `hz` in
+    JzCzhz.
 -   **FIX**: The oRGB color space should be gamut mapped in `srgb` as it is a transform of the RGB space.
 -   **FIX**: Okhsl and Okhsv have a rough sRGB approximation and are instead gamut mapped to their own gamut.
 -   **FIX**: Much more accurate ICtCp matrices.